[Refactor] New package branding: net.ccscript.axfr4azuredns
