+++ checkstyle.xml
@@ -86,7 +86,7 @@
     <!-- Imports -->
     <module name="IllegalImport" />
     <module name="ImportOrder">
-      <property name="groups" value="/^java\./,javax,org"/>
+      <property name="groups" value="/^java\./,javax,org,com,net"/>
       <property name="ordered" value="true"/>
       <property name="separated" value="true"/>
       <property name="option" value="above"/>
