+++ pom.xml
@@ -1,7 +1,7 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <modelVersion>4.0.0</modelVersion>
-  <groupId>net.ccscript.dnsaxfr2azuredns</groupId>
-  <artifactId>dnsaxfr2azuredns</artifactId>
+  <groupId>net.ccscript.axfr4azuredns</groupId>
+  <artifactId>axfr4azuredns</artifactId>
   <version>0.1.0-SNAPSHOT</version>
   <properties>
     <maven.compiler.source>21</maven.compiler.source>
@@ -219,7 +219,7 @@
             <manifest>
               <addClasspath>true</addClasspath>
               <classpathPrefix>lib/</classpathPrefix>
-              <mainClass>net.ccscript.dnsaxfr2azuredns.DNSServerApp</mainClass>
+              <mainClass>net.ccscript.axfr4azuredns.DNSServerApp</mainClass>
             </manifest>
           </archive>
         </configuration>
