+++ src/test/java/net/ccscript/axfr4azuredns/server/configuration/DNSServerConfigurationFactoryTest.java
@@ -1,4 +1,4 @@
-package net.ccscript.dnsaxfr2azuredns.server;
+package net.ccscript.axfr4azuredns.server.configuration;
 
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
@@ -12,15 +12,12 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import org.junit.jupiter.api.Test;
 
-import net.ccscript.dnsaxfr2azuredns.server.configuration.DNSServerConfiguration;
-import net.ccscript.dnsaxfr2azuredns.server.configuration.DNSServerConfiguration.AzureCredentials;
-import net.ccscript.dnsaxfr2azuredns.server.configuration.DNSServerConfiguration.AzureDomain;
-import net.ccscript.dnsaxfr2azuredns.server.configuration.DNSServerConfiguration.DNSDomain;
-import net.ccscript.dnsaxfr2azuredns.server.configuration.DNSServerConfiguration.Server;
-import net.ccscript.dnsaxfr2azuredns.server.configuration.DNSServerConfiguration.Zone;
-import net.ccscript.dnsaxfr2azuredns.server.configuration.DNSServerConfiguration.ZoneTransferType;
-import net.ccscript.dnsaxfr2azuredns.server.configuration.DNSServerConfigurationException;
-import net.ccscript.dnsaxfr2azuredns.server.configuration.DNSServerConfigurationFactory;
+import net.ccscript.axfr4azuredns.server.configuration.DNSServerConfiguration.AzureCredentials;
+import net.ccscript.axfr4azuredns.server.configuration.DNSServerConfiguration.AzureDomain;
+import net.ccscript.axfr4azuredns.server.configuration.DNSServerConfiguration.DNSDomain;
+import net.ccscript.axfr4azuredns.server.configuration.DNSServerConfiguration.Server;
+import net.ccscript.axfr4azuredns.server.configuration.DNSServerConfiguration.Zone;
+import net.ccscript.axfr4azuredns.server.configuration.DNSServerConfiguration.ZoneTransferType;
 
 public class DNSServerConfigurationFactoryTest {
